Add files via upload
Copy the code from code.java file and insert in your arduino uno.
Thanks, Have a nice day


An Arduino Bluetooth car is a remote-controlled vehicle that utilizes an Arduino microcontroller board and Bluetooth technology to establish a wireless connection between the car and a controlling device, such as a smartphone or a computer. It combines the flexibility and programmability of Arduino with the convenience of Bluetooth communication.

Here's a description of the components and functionalities typically found in an Arduino Bluetooth car:

Arduino Board: The heart of the project is an Arduino microcontroller board like Arduino Uno or Arduino Nano. It acts as the control center, receiving commands from the controlling device and executing them to control the car's movement.

Motor Drivers: The Arduino Bluetooth car usually incorporates motor drivers or motor control modules to control the movement of the motors. Motor drivers interface with the Arduino board and enable it to control the speed and direction of the car's motors.

Motors: The car typically consists of two or more DC motors responsible for driving the wheels. These motors receive commands from the Arduino board through the motor drivers and determine the car's speed and direction.

Bluetooth Module: An additional Bluetooth module, such as HC-05 or HC-06, is connected to the Arduino board. It enables wireless communication between the car and the controlling device. The Bluetooth module receives commands from the controlling device and relays them to the Arduino board for execution.

Power Supply: The car requires a suitable power supply, usually a battery pack or a set of rechargeable batteries, to provide power to the Arduino board, motor drivers, and motors.

Controlling Device: The controlling device, such as a smartphone or a computer, communicates with the Arduino Bluetooth car using a Bluetooth connection. A dedicated mobile application or software program is used on the controlling device to send commands for controlling the car's movement.

Control Functions: The Arduino code running on the microcontroller board interprets the commands received from the controlling device and translates them into appropriate actions. The control functions can include moving forward, backward, turning left or right, stopping, and controlling the speed of the motors.

Additional Features: Depending on the project's complexity, an Arduino Bluetooth car can be equipped with additional features such as obstacle detection using ultrasonic sensors, line following using infrared sensors, LED lights, sound modules, and more.

Overall, an Arduino Bluetooth car provides an interactive and fun way to learn about robotics, programming, and wireless communication. It allows users to remotely control a vehicle using their smartphones or computers and can serve as a starting point for exploring more advanced robotics projects.